backend/requirements.txt
fastapi[standard]
uvicorn[standard]>=0.32.0
langchain>=0.3.0
langchain-openai>=0.2.0
pydantic>=2.9.2
pydantic-settings>=2.0.0
html2text>=2024.2.26

# Google API dependencies for Gmail integration
google-auth>=2.0.0
google-auth-oauthlib>=1.0.0
google-auth-httplib2>=0.1.0
google-api-python-client>=2.0.0

# Push notification dependencies
pywebpush>=2.0.0
cryptography>=42.0.0

# Vector store dependencies
qdrant-client>=1.7.0
tiktoken>=0.5.0

# Supabase for database and auth
supabase>=2.0.0
python-jose[cryptography]>=3.3.0
